se agregaron métodos a las pruebas referentes a fondo

+++ src/main/java/uacm/edu/mx/model/Direccion.java
@@ -25,7 +25,7 @@ public class Direccion {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id")
-	private int idDireccion;
+	private Long idDireccion;
 	
 	@NotBlank(message = "calle  es requerido")
 	@Column(name = "calle",nullable = false, length = 50)
